Fix linters error
