+++ src/redux/actionTypes.js
@@ -1,12 +1,3 @@
-import { combineReducers, configureStore } from '@reduxjs/toolkit';
-import thunk from 'redux-thunk';
-import greeting from './greeting/greeting';
+const GET_RANDOM_GREETING = 'rails-react-app/GET_RANDOM_GREETING';
 
-const rootReducer = combineReducers({ greeting });
-
-const store = configureStore({
-  reducer: rootReducer,
-  middleware: [thunk],
-});
-
-export default store;
+export default GET_RANDOM_GREETING;
