+++ src/redux/configureStore.js
@@ -1,3 +1,12 @@
-const GET_RANDOM_GREETING = 'rails-react-app/GET_RANDOM_GREETING';
+import { combineReducers, configureStore } from '@reduxjs/toolkit';
+import thunk from 'redux-thunk';
+import greeting from './greeting/greeting';
 
-export default GET_RANDOM_GREETING;
+const rootReducer = combineReducers({ greeting });
+
+const store = configureStore({
+  reducer: rootReducer,
+  middleware: [thunk],
+});
+
+export default store;
